split clf fitting into new method
There are two lines where a classifier is fitted.
This commit creates a new method that performs the fitting, which can
help to add functionality to the fitting step in the future.

The initial idea was for `_fit_clf` to be a normal method. However, in
that case it could not be called as `self._fit_clf` inside
`_get_feature_shap_values_per_fold` because it is a static method and it
does not recognize `self`.
This might not be the best approach, since, for example, when
implementing the Early Stopping version of ShapRFECV, we might have some
arguments inside `self` which we want to access.