Weekly cronjob no longer tests notebooks (for now) (#265)
As of right now, our unit tests for notebooks are unreliable and as a
result we get false positives. Therefore it is proposed to disable it
for now until we have a better solution how to fix the notebook tests
(which are a bit more extensive examples than normal unit test).